Splitting up the console command from a base application object.
