+++ src/Command/Changelog.php
@@ -1,7 +1,6 @@
 <?php
 namespace Mill\Command;
 
-use Mill\Application;
 use Mill\Config;
 use Mill\Compiler;
 use Symfony\Component\Console\Input\InputArgument;
@@ -9,7 +8,7 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class Changelog extends Application
+class Changelog extends \Mill\Command
 {
     /**
      * @return void
